Persiste atualizado para criar o banco de dados e as tabelas do projeto

+++ Views/pessoa.index.php
@@ -12,13 +12,16 @@ spl_autoload_register(function ($class_name) {
 	<?php
 	use Db\Persiste;
 	use Models\Pessoa;
-	$pessoas = Persiste::GetAllPessoa();
-
-	foreach($pessoas as $p){
+	$persiste = new Persiste();
+	$pessoas = $persiste->GetAllPessoa();
+	if(! is_null($pessoas)) {
+		foreach($pessoas as $p){
 		echo "<tr><td>$p->getid</td><td>$p->getnome</td><td>$p->gettelefone</td>"
 			."<td><a href='pessoa.edit.php?id=$p->getid' class='btn btn-primary btn-small'>Editar</a></td>"
 			."<td><a href='pessoa.delete.php?id=$p->getid' class='btn btn-primary btn-small'>Excluir</a></td></tr>";
+		}	
 	}
+	
 	?>
 	</table>
